move interface to new package and export the package remove the fixed generator option from the configuration include type bump bundle version to 3.3.0
